+++ src/scenes/sceneMesh.cpp
@@ -10,12 +10,12 @@ namespace phi
         auto indices = _mesh->getIndices();
         auto positions = _mesh->getPositionsBuffer();
         auto vertexArray = new btTriangleIndexVertexArray(
-            indices->size() / 3,
+            (int)indices->size() / 3,
             reinterpret_cast<int*>(indices->data()),
-            sizeof(GLuint) * 3,
-            mesh->getVertices().size(),
+            (int)sizeof(GLuint) * 3,
+            (int)mesh->getVertices().size(),
             positions,
-            sizeof(GLfloat) * 3);
+            (int)sizeof(GLfloat) * 3);
 
         _bulletShape = new btGImpactMeshShape(vertexArray);
         _bulletShape->setMargin(btScalar(0.0f));
